requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
redis
prometheus-fastapi-instrumentator
tenacity
python-dotenv
pandas
elasticsearch
python-multipart
playwright
aiogram>=3.0.0
gspread
google-auth
google-auth-oauthlib
pydantic-settings
lxml
